+++ crates/ryo3-twox-hash/src/xxhash32.rs
@@ -3,10 +3,12 @@ use std::hash::Hasher;
 use pyo3::intern;
 use pyo3::prelude::*;
 use pyo3::types::PyString;
-use ryo3_bytes::PyBytes as RyBytes;
-use ryo3_core::RyMutex;
+use ryo3_bytes::ReadableBuffer;
 use ryo3_core::types::{PyDigest, PyHexDigest};
+use ryo3_core::{PyAsciiString, RyMutex};
 use twox_hash::XxHash32;
+
+const HASHLIB_GIL_MINSIZE: usize = 2048;
 
 #[pyclass(name = "xxh32", frozen, immutable_type, skip_from_py_object)]
 #[cfg_attr(feature = "ry", pyo3(module = "ry.ryo3.xxhash"))]
@@ -19,8 +21,7 @@ pub struct PyXxHash32 {
 impl PyXxHash32 {
     #[new]
     #[pyo3(signature = (data = None, *, seed = 0))]
-    fn py_new(data: Option<RyBytes>, seed: Option<u32>) -> Self {
-        let seed = seed.unwrap_or(0);
+    fn py_new(data: Option<ReadableBuffer>, seed: u32) -> Self {
         let hasher = XxHash32::with_seed(seed);
         match data {
             Some(s) => {
@@ -38,10 +39,10 @@ impl PyXxHash32 {
         }
     }
 
-    fn __repr__(&self) -> PyResult<String> {
+    fn __repr__(&self) -> PyResult<PyAsciiString> {
         self.hasher
             .py_lock()
-            .map(|hasher| format!("xxh32<{:x}>", hasher.finish()))
+            .map(|hasher| format!("xxh32<{:x}>", hasher.finish()).into())
     }
 
     #[classattr]
@@ -64,6 +65,11 @@ impl PyXxHash32 {
         self.seed
     }
 
+    #[getter]
+    fn length(&self) -> PyResult<u64> {
+        self.hasher.py_lock().map(|h| h.total_len())
+    }
+
     fn digest(&self) -> PyResult<PyDigest<u32>> {
         let digest = self.hasher.py_lock().map(|h| h.finish_32())?;
 
@@ -80,12 +86,19 @@ impl PyXxHash32 {
     }
 
     #[expect(clippy::needless_pass_by_value)]
-    fn update(&self, py: Python<'_>, data: RyBytes) -> PyResult<()> {
-        py.detach(|| {
+    fn update(&self, py: Python<'_>, data: ReadableBuffer) -> PyResult<()> {
+        let slice = data.as_ref();
+        if slice.len() > HASHLIB_GIL_MINSIZE {
+            py.detach(|| {
+                let mut hasher = self.hasher.py_lock()?;
+                hasher.write(slice);
+                Ok(())
+            })
+        } else {
             let mut hasher = self.hasher.py_lock()?;
-            hasher.write(data.as_ref());
+            hasher.write(slice);
             Ok(())
-        })
+        }
     }
 
     fn copy(&self) -> PyResult<Self> {
@@ -95,6 +108,7 @@ impl PyXxHash32 {
             seed: self.seed,
         })
     }
+
     #[pyo3(signature = (*, seed = None))]
     fn reset(&self, seed: Option<u32>) -> PyResult<()> {
         let mut h = self.hasher.py_lock()?;
@@ -105,8 +119,13 @@ impl PyXxHash32 {
     #[expect(clippy::needless_pass_by_value)]
     #[staticmethod]
     #[pyo3(signature = (data, *, seed = 0))]
-    fn oneshot(py: Python<'_>, data: RyBytes, seed: u32) -> u32 {
-        py.detach(|| XxHash32::oneshot(seed, data.as_ref()))
+    fn oneshot(py: Python<'_>, data: ReadableBuffer, seed: u32) -> PyDigest<u32> {
+        let slice = data.as_ref();
+        if slice.len() > HASHLIB_GIL_MINSIZE {
+            py.detach(|| XxHash32::oneshot(seed, slice).into())
+        } else {
+            XxHash32::oneshot(seed, slice).into()
+        }
     }
 }
 
@@ -116,22 +135,37 @@ impl PyXxHash32 {
 #[expect(clippy::needless_pass_by_value)]
 #[pyfunction]
 #[pyo3(signature = (data, *, seed = 0))]
-pub fn xxh32_digest(py: Python<'_>, data: RyBytes, seed: u32) -> PyDigest<u32> {
-    py.detach(|| twox_hash::XxHash32::oneshot(seed, data.as_ref()).into())
+pub fn xxh32_digest(py: Python<'_>, data: ReadableBuffer, seed: u32) -> PyDigest<u32> {
+    let slice = data.as_ref();
+    if slice.len() > HASHLIB_GIL_MINSIZE {
+        py.detach(|| XxHash32::oneshot(seed, slice).into())
+    } else {
+        XxHash32::oneshot(seed, slice).into()
+    }
 }
 
 #[expect(clippy::needless_pass_by_value)]
 #[pyfunction]
 #[pyo3(signature = (data, *, seed = 0))]
-pub fn xxh32_intdigest(py: Python<'_>, data: RyBytes, seed: u32) -> u32 {
-    py.detach(|| twox_hash::XxHash32::oneshot(seed, data.as_ref()))
+pub fn xxh32_intdigest(py: Python<'_>, data: ReadableBuffer, seed: u32) -> u32 {
+    let slice = data.as_ref();
+    if slice.len() > HASHLIB_GIL_MINSIZE {
+        py.detach(|| XxHash32::oneshot(seed, slice))
+    } else {
+        XxHash32::oneshot(seed, slice)
+    }
 }
 
 #[expect(clippy::needless_pass_by_value)]
 #[pyfunction]
 #[pyo3(signature = (data, *, seed = 0))]
-pub fn xxh32_hexdigest(py: Python<'_>, data: RyBytes, seed: u32) -> PyHexDigest<u32> {
-    py.detach(|| twox_hash::XxHash32::oneshot(seed, data.as_ref()).into())
+pub fn xxh32_hexdigest(py: Python<'_>, data: ReadableBuffer, seed: u32) -> PyHexDigest<u32> {
+    let slice = data.as_ref();
+    if slice.len() > HASHLIB_GIL_MINSIZE {
+        py.detach(|| XxHash32::oneshot(seed, slice).into())
+    } else {
+        XxHash32::oneshot(seed, slice).into()
+    }
 }
 
 pub fn pymod_add(m: &Bound<'_, PyModule>) -> PyResult<()> {