+++ tests/reqwest/conftest.py
@@ -21,6 +21,7 @@ from uvicorn.server import Server
 
 import ry
 
+TClient: t.TypeAlias = ry.HttpClient | ry.Client | ry.BlockingClient  # type: ignore[deprecated]
 Receive: t.TypeAlias = t.Callable[[], Awaitable[dict[str, t.Any]]]
 Send: t.TypeAlias = t.Callable[[uvt.ASGISendEvent], Coroutine[None, None, None]]
 Scope: t.TypeAlias = dict[str, t.Any]
@@ -352,11 +353,9 @@ def server() -> Iterator[ReqtestServer]:
         yield running
 
 
-@pytest.fixture(params=[ry.HttpClient, ry.Client, ry.BlockingClient])
-def client_cls(
-    request: pytest.FixtureRequest,
-) -> type[ry.HttpClient | ry.Client | ry.BlockingClient]:
-    return t.cast("type[ry.HttpClient | ry.Client | ry.BlockingClient]", request.param)
+@pytest.fixture(params=[ry.HttpClient, ry.Client, ry.BlockingClient])  # type: ignore[deprecated]
+def client_cls(request: pytest.FixtureRequest) -> type[TClient]:
+    return t.cast("type[TClient]", request.param)
 
 
 def _main() -> None: